+++ R/utils.R
@@ -1,4 +1,4 @@
-#' @importFrom rlang abort arg_match is_named
+#' @importFrom rlang abort arg_match is_named %||%
 # Include generic helpers from ggplot2
 #' @include utils_ggplot2.R
 #' @include utils_ggplot2_performance.R